Add lemma_encoding=2 to morphodita_parsito model and make it default.
The new lemma encoding model can handle the situation when
only *some* lemmas are empty (either "" or "_"), while others are not.
This is required for UD 1.3 fa corpus which fills only some lemmas.

The new lemma encoding can be used to also encode a fixed set of
"absolute" lemmas (i.e., not generated from the form), which is useful
for lemma greek.expression in la_proiel corpus.