Add input_format::reset_document call.
The call resets the input_format instance state. Such state is needed
not only for remembering unprocessed text of the last set_text call, but
also for correct inter-block state tracking (for example in a tokenizer
which returns document-level offsets, or in a tokenizer which tracks
inter-sentence spaces -- if you pass only spaces to set_text, these
spaces has to accumulate and be returned as preceeding spaces of the
next sentence).