Rething form and lemma normalization, bump model_morphodita_parsito version to 3.
We now normalize forms even for the parser, replacing ' ' by a &nbsp; to
allow word embeddings with spaces. We also normalize lemmas in the same
way.

We need model version bump, as the change is backward incompatible
(and we have already updated the GRU tokenizer). Note that more changes
in the version 3 serialization will come.