Rename 'dictionary' tagger option to 'dictionary_file'.
Instead of previous non-intuitive (and less efficient)
  dictionary=file:f
we now use
  dictionary_file=f
The change requires the dictionary to be in a file (and not just in
memory), but that is the case almost every time.

The old 'dictionary' option now generates an error during training,
pointing to the new 'dictionary_file' option.

Fixes #36.