Only load the model networks when needed.
Specifying --preload_models loads all networks on startup.

Loading happens in a global mutex so that parallel requests
do not load the network twice. The mutex could be per-network,
but using a global one seems reasonable anyway.

Even if the networks are not preloaded, tokenizers currently
always are. But it is fast and does not take much memory.